pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
  "Natural Language :: English",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
name = "pyssmf"
description = "A Python package to run Slave-Spin Mean-Field (SSMF) simulations in strongly correlated electron systems."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Youssra Anene", email = "youssraanene20@gmail.com" },
  { name = "Jose M. Pizarro", email = "jose.pizarro-blanco@bam.de" },
  { name = "Lorenzo Fratino" },
]
maintainers = [
  { name = "Youssra Anene", email = "youssraanene20@gmail.com" },
  { name = "Jose M. Pizarro", email = "jose.pizarro-blanco@bam.de" },
]
license = { file = "LICENSE" }
dependencies = [
  "numpy",
  "scipy",
  "pydantic~=2.10.5",
  "pyyaml",
  "h5py~=3.12.1",
  "structlog==24.4.0",
]

[project.urls]
"Homepage" = "https://github.com/Yanene/pySSMF"
"Bug Tracker" = "https://github.com/Yanene/pySSMF/issues"
"Documentation" = "https://yanene.github.io/pySSMF/"

[project.optional-dependencies]
dev = [
  "mypy==1.0.1",
  "ruff",
  "pytest",
  "pytest-timeout",
  "pytest-cov",
  "pre-commit",
  "detect-secrets",
]
docu = [
  "mkdocs-material",
  "mkdocs-click",
  "mkdocstrings",
  "mkdocstrings-python",
]

[tool.uv]
build-constraint-dependencies = ["wheel<0.46"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/pyssmf/_version.py"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
include = ["src/pyssmf/*.py", "tests/*.py"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "dependencies",
    "scripts"
]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "PL", # pylint
    "F", # Pyflakes
    "UP", # pyupgrade
    "I", # isort
]
ignore = [
    "F401", # Module imported but unused
    "E501", # Line too long ({width} > {limit} characters)
    "E701", # Multiple statements on one line (colon)
    "E731", # Do not assign a lambda expression, use a def
    "E402",  # Module level import not at top of file
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used instead of constant
    "PLW0603", # Using the global statement
    "PLW2901", # redefined-loop-name
    "PLR1714", # consider-using-in
    "PLR5501", # else-if-used
]
fixable = ["ALL"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use double quotes for strings.
quote-style = "double"
# indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
